+++ client/src/game/tools/variants/spell.ts
@@ -152,7 +152,7 @@ class SpellTool extends Tool implements ITool {
         accessSystem.addAccess(
             this.shape.id,
             playerSystem.getCurrentPlayer()!.name,
-            { edit: true, movement: true, vision: true },
+            { edit: true, movement: true, vision: false },
             UI_SYNC,
         );